Kill tidying logic, move DSL functions to `test_ki.py`
